host/requirements.txt
websockets>=12.0
aiohttp>=3.9
pyserial-asyncio>=0.6
